# src/swagger.json
{
    "components": {
...
        "securitySchemes": {}
    },
    "info": {
        "title": "Technik Informatyk API",
        "version": "v1",
        "description": "REST API for Technik Informatyk App",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/question/{uuid}/image": {
            "get": {
                "operationId": "GetQuestionImage",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "enum": ["Not found", "Not Found"]
                                }
                            }
                        }
                    }
                },
                "tags": ["Question"],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}